renamed test stats dir name
