+++ .gitignore
@@ -11,3 +11,4 @@ maintainernotes.md
 tutorials/auspopn.zip
 tutorials/analysis_grid.nc
 tutorials/forecast_grid.nc
+report.xml
