+++ src/scores/stats/__init__.py
@@ -1,4 +1,4 @@
 """
 The philosphy is to import the public API during the init phase rather than leaving it to the user
 """
-import scores.stats.tests
+import scores.stats.statistical_tests
