+++ src/scores/stats/statistical_tests/__init__.py
@@ -1,0 +1,5 @@
+"""
+Import the functions from the implementations into the public API
+"""
+
+from scores.stats.statistical_tests.diebold_mariano_impl import diebold_mariano
