--- src/scores/stats/tests/__init__.py
@@ -1,5 +1,0 @@
-"""
-Import the functions from the implementations into the public API
-"""
-
-from scores.stats.tests.diebold_mariano_impl import diebold_mariano
