+++ tests/stats/test_diebold_mariano.py
@@ -7,7 +7,7 @@ import numpy as np
 import pytest
 import xarray as xr
 
-from scores.stats.tests.diebold_mariano_impl import (
+from scores.stats.statistical_tests.diebold_mariano_impl import (
     _dm_gamma_hat_k,
     _dm_test_statistic,
     _dm_v_hat,
